+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `HSIButton`: Allow setting the feature_count property via prop.
+- `Footer`: Allow passing the `className` prop.
 
 ## [2.0.0] - 2021-07-20